scripts: move old tools directory's content
