+++ demo/standalone/server.js
@@ -4,16 +4,16 @@ const Webpack = require("webpack");
 const WebpackDevServer = require("webpack-dev-server");
 const path = require("path");
 
-const webpackConfig = require("./webpack.config.js");
+const webpackConfig = require("../../webpack.config.js");
 
 // overwrite entries/output (ugly but just werks and did not find any better)
-webpackConfig.entry = path.join(__dirname, "../src/exports.ts");
+webpackConfig.entry = path.join(__dirname, "../../src/exports.ts");
 webpackConfig.output.path = path.join(__dirname, "../demo/standalone");
 webpackConfig.output.filename = "lib.js";
 
 const compiler = Webpack(webpackConfig);
 const server = new WebpackDevServer(compiler, {
-  contentBase: path.join(__dirname, "../demo/standalone"),
+  contentBase: path.join(__dirname, "./"),
   compress: true,
 });
