+++ scripts/doc-generator/construct_table_of_contents.js
