+++ scripts/doc-generator/create_for_dir.js
