+++ scripts/doc-generator/get_files_to_convert.js
