+++ scripts/doc-generator/index.js
