+++ scripts/doc-generator/mkdir_parent.js
